# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Goaldone API
  description: |
    **Goaldone** – Browserbasiertes Aufgaben-Planungssystem für Organisationen (TINF2024 Projektarbeit).

    ## Authentifizierung
    Die API verwendet JWT Bearer Tokens kombiniert mit einem HttpOnly Cookie für den Refresh Token.
    - **Access Token**: kurzlebig (15 Min.), für jeden API-Aufruf als `Authorization: Bearer <token>` Header mitzusenden
    - **Refresh Token**: langlebig (7 Tage), wird ausschließlich als HttpOnly Cookie (`refresh_token`) übertragen.
      JavaScript-Code kann diesen Token niemals lesen (XSS-Schutz).
      Der Cookie wird nur an `/auth/refresh` und `/auth/logout` gesendet (`Path=/api/v1/auth`, `SameSite=Strict`).
    - Token Rotation: Bei jedem Refresh wird ein neues Refresh Token Cookie ausgestellt, das alte wird invalidiert

    Der JWT enthält folgende Claims:
    - `sub` – UUID des Nutzers
    - `organizationId` – UUID der Organisation (null bei SUPER_ADMIN ohne Org)
    - `role` – Rolle des Nutzers (`SUPER_ADMIN`, `ADMIN`, `USER`)

    ## Rollen (RBAC)
    | Rolle | Beschreibung |
    |---|---|
    | `SUPER_ADMIN` | Plattformweite Verwaltung (Organisationen, Super-Admins) |
    | `ADMIN` | Verwaltung der **eigenen** Organisation (Mitglieder, Einladungen, Einstellungen) |
    | `USER` | Standard-Nutzer mit Zugriff auf eigene Aufgaben und Planung |

    ## Mandantenfähigkeit & Tenant-Isolation
    Goaldone ist eine **mandantenfähige Plattform** (Multi-Tenant). Jede Ressource
...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/refresh:
    post:
      tags: [auth]
      summary: Access Token erneuern (Token Rotation)
      description: |
        Liest den Refresh Token aus dem HttpOnly Cookie `refresh_token`.
        Gibt einen neuen Access Token im Body zurück und setzt einen neuen Refresh Token Cookie.
        Der alte Refresh Token Cookie wird serverseitig invalidiert (Token Rotation).
        Kein Request Body erforderlich – der Cookie wird vom Browser automatisch mitgesendet.
      security: []
      operationId: refreshToken
      parameters:
        - in: cookie
          name: refresh_token
          required: true
          schema:
            type: string
          description: "HttpOnly Refresh Token Cookie (wird vom Browser automatisch mitgesendet)"
      responses:
        '200':
          description: Token erfolgreich erneuert
          headers:
            Set-Cookie:
              description: "Neuer Refresh Token Cookie (Token Rotation, ersetzt den alten)"
              schema:
                type: string
                example: "refresh_token=eyJ...; HttpOnly; Secure; SameSite=Strict; Path=/api/v1/auth; Max-Age=604800"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /tasks:
    get:
      tags: [tasks]
      summary: Eigene Tasks auflisten
      description: |
        Gibt paginiert alle Tasks des eingeloggten Nutzers zurück.
        Filterbar nach Status und Deadline-Zeitraum.
      operationId: listTasks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: status
          in: query
          description: Nur Tasks mit diesem Status zurückgeben
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: from
          in: query
          description: Nur Tasks mit Deadline ab diesem Datum (ISO 8601 Date)
          schema:
            type: string
            format: date
            example: "2026-04-01"
        - name: to
          in: query
          description: Nur Tasks mit Deadline bis zu diesem Datum (ISO 8601 Date)
          schema:
            type: string
            format: date
            example: "2026-04-30"
      responses:
        '200':
          description: Paginierte Task-Liste
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskPage'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [tasks]
      summary: Neuen Task erstellen
      description: |
        Erstellt einen neuen Task für den eingeloggten Nutzer.
        Wenn `estimatedDurationMinutes` > 480 (8h), wird empfohlen, den Task zu splitten
        – dies kann entweder clientseitig angeboten oder serverseitig automatisch erfolgen.

        **Einschränkung:** `SUPER_ADMIN`-Nutzer dürfen keine Tasks erstellen (`403 Forbidden`).
        Tasks können ausschließlich von `USER`- und `ADMIN`-Nutzern angelegt werden.
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task erfolgreich erstellt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: SUPER_ADMIN darf keine Tasks erstellen
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: "https://goaldone.de/problems/forbidden"
                title: "Forbidden"
                status: 403
                detail: "super-admin-cannot-create-tasks"
                instance: "/api/v1/tasks"
  /tasks/{taskId}:
    get:
      tags: [tasks]
      summary: Einzelnen Task abrufen
      operationId: getTask
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '200':
          description: Task-Details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [tasks]
      summary: Task vollständig aktualisieren
      operationId: updateTask
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [tasks]
      summary: Task löschen (Hard Delete)
      operationId: deleteTask
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '204':
          description: Task gelöscht
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks/{taskId}/complete:
    patch:
      tags: [tasks]
      summary: Task als erledigt markieren
      description: Setzt den Status des Tasks auf DONE und setzt completedAt.
      operationId: completeTask
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '200':
          description: Task als erledigt markiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Task ist bereits erledigt
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /tasks/{taskId}/reopen:
    patch:
      tags: [tasks]
      summary: Erledigten Task wieder öffnen
      description: Setzt den Status des Tasks von DONE zurück auf OPEN.
      operationId: reopenTask
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '200':
          description: Task wieder geöffnet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Task ist nicht erledigt
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  # ════════════════════════════════════════════
  # BREAKS
  # ════════════════════════════════════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Access Token. Format: Authorization: Bearer <token>"

    cookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: "HttpOnly Refresh Token Cookie. Wird ausschließlich von /auth/refresh und /auth/logout verwendet."
  parameters:
    PageParam:
      name: page
      in: query
      description: Seitennummer (0-basiert)
      schema:
        type: integer
        default: 0
        minimum: 0
      example: 0
    SizeParam:
      name: size
      in: query
      description: Anzahl Einträge pro Seite
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      example: 20
    TaskIdParam:
      name: taskId
      in: path
      required: true
      description: UUID des Tasks
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Ungültige Anfrage (Validierungsfehler)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: "https://goaldone.de/problems/validation-error"
            title: "Validation Error"
            status: 400
            detail: "Das Feld 'title' darf nicht leer sein."
            instance: "/api/v1/tasks"
    Unauthorized:
      description: Nicht authentifiziert – JWT fehlt oder ist ungültig/abgelaufen
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: "https://goaldone.de/problems/unauthorized"
            title: "Unauthorized"
            status: 401
            detail: "JWT Token ist abgelaufen."
            instance: "/api/v1/tasks"
    Forbidden:
      description: Keine Berechtigung – Rolle reicht nicht aus
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: "https://goaldone.de/problems/forbidden"
            title: "Forbidden"
            status: 403
            detail: "Diese Aktion erfordert die Rolle ADMIN."
            instance: "/api/v1/organizations/me/members"
    NotFound:
      description: Ressource nicht gefunden
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: "https://goaldone.de/problems/not-found"
            title: "Not Found"
            status: 404
            detail: "Task mit ID 'abc123' wurde nicht gefunden."
            instance: "/api/v1/tasks/abc123"
...
  schemas:
...
    ProblemDetail:
      type: object
      description: Fehlerantwort nach RFC 9457 (Problem Details for HTTP APIs)
      properties:
        type:
          type: string
          format: uri
          description: URI, die den Fehlertyp eindeutig identifiziert
          example: "https://goaldone.de/problems/validation-error"
        title:
          type: string
          description: Kurze, menschenlesbare Fehlerbeschreibung
          example: "Validation Error"
        status:
          type: integer
          description: HTTP Status Code
          example: 400
        detail:
          type: string
          description: Detaillierte Fehlermeldung für diesen spezifischen Aufruf
          example: "Das Feld 'title' darf nicht leer sein."
        instance:
          type: string
          format: uri
          description: URI des Endpunkts, der den Fehler verursacht hat
          example: "/api/v1/tasks"
        errors:
          type: array
          description: Feldspezifische Validierungsfehler (nur bei 400 Bad Request)
          items:
            $ref: '#/components/schemas/FieldError'
      required: [type, title, status, detail, instance]
    FieldError:
      type: object
      description: Validierungsfehler für ein einzelnes Feld
      properties:
        field:
          type: string
          example: "title"
        message:
          type: string
          example: "Darf nicht leer sein"
      required: [field, message]

    # ── Auth ──────────────────────────────────────────────
...
    RefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: "Neuer JWT Access Token, gültig 15 Minuten"
      required: [accessToken]
...
    TaskPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/TaskResponse'
          required: [content]

    # ── Task ──────────────────────────────────────────────
    TaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/TaskStatus'
        cognitiveLoad:
          $ref: '#/components/schemas/CognitiveLoad'
        estimatedDurationMinutes:
          type: integer
          description: "Geschätzte Arbeitsdauer in Minuten"
          example: 90
        deadline:
          type: string
          format: date
          nullable: true
        recurrence:
          $ref: '#/components/schemas/RecurrenceRule'
          nullable: true
        parentTaskId:
          type: string
          format: uuid
          nullable: true
          description: "Gesetzt wenn dieser Task durch automatisches Task-Splitting entstanden ist (Ursprungs-Task-Referenz)"
        ownerId:
          type: string
          format: uuid
          description: "Aktuell immer der eingeloggte Nutzer. Vorbereitet für zukünftige Team-Features."
        organizationId:
          type: string
          format: uuid
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, status, cognitiveLoad, estimatedDurationMinutes, ownerId, organizationId, createdAt, updatedAt]
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          example: "Anforderungsanalyse fertigstellen"
        description:
          type: string
          nullable: true
          maxLength: 2000
        cognitiveLoad:
          $ref: '#/components/schemas/CognitiveLoad'
        estimatedDurationMinutes:
          type: integer
          minimum: 1
          example: 90
          description: "Geschätzte Arbeitsdauer in Minuten. Bei > 480 (8h) empfiehlt das System Splitting."
        deadline:
          type: string
          format: date
          nullable: true
          example: "2026-04-30"
        recurrence:
          $ref: '#/components/schemas/RecurrenceRule'
          nullable: true
      required: [title, cognitiveLoad, estimatedDurationMinutes]
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
          maxLength: 2000
        cognitiveLoad:
          $ref: '#/components/schemas/CognitiveLoad'
        estimatedDurationMinutes:
          type: integer
          minimum: 1
        deadline:
          type: string
          format: date
          nullable: true
        recurrence:
          $ref: '#/components/schemas/RecurrenceRule'
          nullable: true
      required: [title, cognitiveLoad, estimatedDurationMinutes]
    RecurrenceRule:
      type: object
      description: |
        Wiederholungsregel für Tasks und Pausen.

        **Aktuelles Modell (MVP):** Typ + Intervall
        Beispiele:
        - `{ "type": "DAILY", "interval": 1 }` → jeden Tag
        - `{ "type": "WEEKLY", "interval": 2 }` → alle 2 Wochen
        - `{ "type": "MONTHLY", "interval": 1 }` → jeden Monat

        **Erweiterbarkeit:** Das Modell ist so strukturiert, dass es in einer
        späteren Version um ein `rrule`-Feld (iCal RRULE-Standard) ergänzt werden
        kann, ohne bestehende Clients zu brechen.
      properties:
        type:
          $ref: '#/components/schemas/RecurrenceType'
        interval:
          type: integer
          minimum: 1
          default: 1
          description: "Alle N Einheiten wiederholen (abhängig von type)"
          example: 1
      required: [type, interval]

    # ── Break ─────────────────────────────────────────────
...
    PageMetadata:
      type: object
      description: "Paginierungsmetadaten, enthalten in allen paginierten Antworten"
      properties:
        page:
          type: integer
          description: "Aktuelle Seitennummer (0-basiert)"
          example: 0
        size:
          type: integer
          description: "Einträge pro Seite"
          example: 20
        totalElements:
          type: integer
          description: "Gesamtanzahl aller Einträge"
          example: 42
        totalPages:
          type: integer
          description: "Gesamtanzahl der Seiten"
          example: 3
      required: [page, size, totalElements, totalPages]

    # ── Enums ─────────────────────────────────────────────
...
    TaskStatus:
      type: string
      enum: [OPEN, IN_PROGRESS, DONE]
      description: |
        Status eines Tasks:
        - `OPEN`: Noch nicht begonnen
        - `IN_PROGRESS`: Wird gerade bearbeitet
        - `DONE`: Abgeschlossen (setzt `completedAt`)
    CognitiveLoad:
      type: string
      enum: [LOW, MEDIUM, HIGH]
      description: |
        Kognitive Belastung der Aufgabe (aus Anforderungsdokument: "Wie fordernd ist die Aufgabe?"):
        - `LOW`: Routine / stupide Aufgabe (z.B. E-Mails sortieren)
        - `MEDIUM`: Normaler Arbeitsaufwand
        - `HIGH`: Höchste Konzentration erforderlich (z.B. komplexe Analyse)

        Der Planungsalgorithmus nutzt dieses Feld um `HIGH`-Tasks bevorzugt in
        produktive Morgenstunden einzuplanen.
    RecurrenceType:
      type: string
      enum: [DAILY, WEEKLY, MONTHLY]
      description: |
        Wiederholungstyp (MVP-Modell, erweiterbar zu RRULE):
        - `DAILY`: Täglich (z.B. E-Mails checken)
        - `WEEKLY`: Wöchentlich
        - `MONTHLY`: Monatlich